step counter page and update
