+++ app/src/main/java/com/personalchef/mealplan/models/StepCounter.java
@@ -1,14 +1,11 @@
 package com.personalchef.mealplan.models;
 
 import android.content.SharedPreferences;
-import android.hardware.Sensor;
-import android.hardware.SensorEvent;
-import android.hardware.SensorEventListener;
-import android.hardware.SensorManager;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.personalchef.mealplan.MainActivity;
+import com.personalchef.mealplan.StepCounterActivity;
 
 import java.util.Date;
 
@@ -33,8 +30,8 @@ public class StepCounter extends AppCompatActivity {
     public int GetStepCount() {
         MainActivity mainActivity = new MainActivity();
         System.out.println("From GetStepCounter\n_____________________________");
-        System.out.println(mainActivity.stepCount);
-        return mainActivity.stepCount;
+        System.out.println(StepCounterActivity.stepCount);
+        return StepCounterActivity.stepCount;
     }
 
     public int  CalculateCaloriesBurnt() {
