+++ app/src/main/res/drawable/stats_progress.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<rotate xmlns:android="http://schemas.android.com/apk/res/android"
+    android:fromDegrees="270"
+    android:toDegrees="270">
+    <shape
+        android:innerRadius="100dp"
+        android:shape="ring"
+        android:thickness="30dp"
+        android:useLevel="true">
+        <solid android:color="@color/colorAccent" />
+    </shape>
+</rotate>
