+++ app/src/main/res/values/colors.xml
@@ -10,4 +10,6 @@
     <color name="white">#FFFFFFFF</color>
     <color name="myorange">#fca103</color>
     <color name="gray1">#9E9A9A</color>
+    <color name="colorPrimaryDark">#9C27B0</color>
+    <color name="colorAccent">#FFEB3B</color>
 </resources>
